Recognizer evolves + Some small TODOs

+++ FaceRecognitionDeployment/fr_project/authentication/models.py
@@ -7,7 +7,6 @@ from django.conf import settings
 import os
 import shutil
 from .faceRecognition import createRecognizer
-# Create your models here.
 
 # Create your models here.
 class User(AbstractBaseUser, PermissionsMixin):
@@ -33,7 +32,7 @@ class User(AbstractBaseUser, PermissionsMixin):
 
     recognizer = models.CharField(
         _("recognizer"),
-        max_length=4096, # TODO: a lo mejor queremos hacer mas pequeño el campo hehe
+        max_length=4096,
         unique=True,
         null =True,
         blank=True,
@@ -112,12 +111,13 @@ class User(AbstractBaseUser, PermissionsMixin):
         file.write(str(numFaces) + '\n')
         file.close
 
-
         return numRequests, numFaces
 
 
     def cleanUserFolder(self):
-
-        shutil.rmtree(self.get_tmp_raw_imgs_path())
-        shutil.rmtree(self.get_tmp_processed_imgs_path())
-        os.remove(self.get_metadata_path())
+        if os.path.isdir(self.get_tmp_raw_imgs_path()):
+            shutil.rmtree(self.get_tmp_raw_imgs_path())
+        if os.path.isdir(self.get_tmp_processed_imgs_path()):
+            shutil.rmtree(self.get_tmp_processed_imgs_path())
+        if os.path.isfile(self.get_metadata_path()):
+            os.remove(self.get_metadata_path())
